cpu_boost: Only restore minfreq to CPU0
Previously, when a non-primary CPU core was boosted, if it was offline after a boost expired, then the minfreq on that CPU would still be set to the boost frequency. Minfreq is now synced across all cores in msm/cpufreq, so we only need to restore the minfreq to CPU0.

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>